# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/models": {
      "get": {
        "operationId": "ModelsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "列出当前用户可用于新会话的模型",
        "tags": [
          "Models"
        ]
      }
    },
    "/api/v1/custom-models": {
      "get": {
        "operationId": "CustomModelsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "列出当前用户自己的自定义模型",
        "tags": [
          "Custom Models"
        ]
      },
      "post": {
        "operationId": "CustomModelsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomModelRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建用户自定义 OpenAI-compatible 模型",
        "tags": [
          "Custom Models"
        ]
      }
    },
    "/api/v1/custom-models/{id}": {
      "patch": {
        "operationId": "CustomModelsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomModelRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建自定义模型的新草稿版本",
        "tags": [
          "Custom Models"
        ]
      },
      "delete": {
        "operationId": "CustomModelsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除自定义模型并立即清除其加密凭据",
        "tags": [
          "Custom Models"
        ]
      }
    },
    "/api/v1/custom-models/{id}/publish": {
      "post": {
        "operationId": "CustomModelsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "发布自定义模型当前草稿版本",
        "tags": [
          "Custom Models"
        ]
      }
    },
    "/api/v1/custom-models/{id}/disable": {
      "post": {
        "operationId": "CustomModelsController_disable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "停用自定义模型",
        "tags": [
          "Custom Models"
        ]
      }
    },
    "/api/v1/custom-models/{id}/test": {
      "post": {
        "operationId": "CustomModelsController_testConnection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConnectionTestResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "测试自定义模型连接，不回显凭据或原始错误体",
        "tags": [
          "Custom Models"
        ]
      }
    },
    "/api/v1/admin/models": {
      "get": {
        "operationId": "AdminModelsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "管理员列出系统模型",
        "tags": [
          "Admin Models"
        ]
      },
      "post": {
        "operationId": "AdminModelsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminModelRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "管理员创建系统模型草稿",
        "tags": [
          "Admin Models"
        ]
      }
    },
...
    "/api/v1/admin/models/{id}/publish": {
      "post": {
        "operationId": "AdminModelsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "发布模型当前草稿版本，仅影响新建会话",
        "tags": [
          "Admin Models"
        ]
      }
    },
    "/api/v1/admin/models/{id}/disable": {
      "post": {
        "operationId": "AdminModelsController_disable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求字段或格式不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或使用了无效身份凭据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行该操作或访问该资源",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "请求超过限流阈值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "停用系统模型，不影响历史会话",
        "tags": [
          "Admin Models"
        ]
      }
    },
...
  },
  "info": {
    "title": "Destiny Compiler API",
    "description": "命运编译器 MVP 代码契约",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "perPage": {
            "type": "number",
            "example": 20,
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "number",
            "example": 42,
            "minimum": 0
          },
          "totalPages": {
            "type": "number",
            "example": 3,
            "minimum": 0
          }
        },
        "required": [
          "page",
          "perPage",
          "total",
          "totalPages"
        ]
      },
      "ModelListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateCustomModelRequestDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 100,
            "example": "我的 API 模型"
          },
          "protocol": {
            "type": "string",
            "enum": [
              "openai_compatible"
            ],
            "example": "openai_compatible"
          },
          "baseUrl": {
            "type": "string",
            "format": "uri",
            "maxLength": 500,
            "example": "https://api.example.com/v1",
            "description": "用户自定义模型首版只接受公网 HTTPS；SSRF 校验由 Service 再执行"
          },
          "modelName": {
            "type": "string",
            "maxLength": 120,
            "example": "example-model"
          },
          "provider": {
            "type": "string",
            "maxLength": 50,
            "example": "openai"
          },
          "apiKey": {
            "type": "string",
            "writeOnly": true,
            "maxLength": 4096
          },
          "timeoutMs": {
            "type": "number",
            "default": 60000,
            "minimum": 1000,
            "maximum": 300000
          },
          "maxOutputTokens": {
            "type": "number",
            "default": 4096,
            "minimum": 1,
            "maximum": 100000
          },
          "supportsStream": {
            "type": "boolean",
            "default": true
          },
          "supportsStructuredOutput": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "displayName",
          "protocol",
          "baseUrl",
          "modelName"
        ]
      },
...
      "ModelResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ModelDetailDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "ModelConnectionTestResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ModelConnectionTestDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}